Day 47 - Challenge 4 - allow incrementing of 'timesCompleted'
We have to convert `Activity` to an `ObservableObject` and also make
sure we have a separate `ActivityRowView` structure, each with its own
`@ObservedObject` reference to an activity.  If we don't do this, then
the rows in the list view don't get notification that the
`timesCompleted` count has incremented, and so won't update.